Initial drop of generated types
This is intended to be a first usable attempt at a Go SDK for Eiffel
events. It covers the events from the Lyon edition.

The principle chosen is to have one Go struct per major version of each
event. This means that v3.0.0 of some event received over the wire might
be deserialized into a struct that supports up to v3.3.0. That version
defines additional fields, which will obviously be empty. Similarly, if
you instantiate the V3 struct for that event type you need to be careful
about which struct fields you set if you don't intend to use the most
recent version within that major version.

Non-exhaustive list of missing features:

    - Factory function for each struct that populates basic required
      fields like meta.type, meta.version, meta.id, and meta.time.
    - JSON serialization that (optionally) includes validation.
    - Subpackages for all Eiffel editions that define type aliases
      that point to concrete structs with major versions matching
      the contents of the edition.

The structs and other types are generated based on the schemas. For that
reason the Eiffel protocol repository is included as a Git submodule.