GitHub actions: Don't install GNU make (#6)
The build-and-test.yml workflow unnecessarily installed GNU make even
though it's already available thanks to the documented inclusion of
the build-essential package for the Ubuntu virtual environment.

This added upwards of 10 s to the workflow execution. While the package
installation itself is a no-op since the package is already installed
the apt-get update operation takes time.