Add modifier support to generated factories
We introduce the event modifier as a function that can be passed
(possibly with other modifiers) to an event factory and will be
called on the newly created event. This can be used to extend the
default population of fields with e.g. meta.source.host and
meta.source.domainId.