Add factory function for each event struct (#12)
The generated factory functions make it easy to create events for
every major event version populated with the mandatory meta fields.
meta.version will be set to the most recent known version (within
that major version). This'll make it much easier to create a complete
and valid event:

    event := NewCompositionDefinedV3()
    event.Data.Name = "my-composition"

A follow-up commit will add support for passing type-agnostic event
modifiers to the factory function, but since this will be done via
variadic arguments the change will be (mostly) backwards compatible.