Generate edition-specific packages
To make it easier to work with the event versions from a particular
Eiffel edition we add subpackages named after the editions that contain
version-less type aliases for each event type and a version-less factory
function that creates events with the correct meta.version.

We had to set fetch-depth to 0 (zero) for the actions/checkout@v2 step
in the build-and-test GitHub actions workflow since the default setting
initialized the submodules with --depth=1, i.e. no tags and no history
were included.