Make the links slice a type of its own with helpers methods
To make it less tedious to write code that links events we introduce
a distinct type of the links field in each event struct. Instead of
having the type []XXXLink we now declare a custom type XXXLinks that
includes a couple of methods for appending links of the right type
to the slice. This allows you to simplify code like

    event.Links = append(event.Links, eiffelevents.ActTV3Link{
            Type:   "CAUSE",
            Target: other.Meta.ID,
    })

to

    event.Links.Add("CAUSE", other)

which not only is a lot shorter but also doesn't include the name of
a versioned type, i.e. edition upgrades are less likely to require
type reference updates.

To support passing an arbitrary event to the Add method without having
to use reflection to locate its meta.id we add a MetaTeller interface
that all events implement. The interface contains methods for obtaining
basic meta fields, including the event's id.