Add LinkFinder interface with FindFirst and FindAll methods
The methods in this interface are added to all XXXLinks types and
they can be used as shorthands to locate links of a particular type
in the full set of links for an event. While the code for doing the
search is trivial it's not reusable between different XXXLinks types
so these methods should be useful to consumers of multiple event
types.