Add modifiers for meta.source.serializer
These new Modifier functions populate meta.source.serializer, either
with an explicitly provided value or automatically based on the
program's main package, e.g. pkg:golang/github.com/a/b@1.0.0. In the
latter case the version needs to be supplied by the caller since it
doesn't appear to be trivial to figure it out automatically (even with
Go 1.18's better VCS support), but that's something we can revisit later
if needed.