internal/cmd/eventgen: Read YAML definition files instead of JSON schemas
We read the new YAML-based schema definition files and adapt the code to
deal with the JSON references that are used in those files. A new goType
type, goPredeclaredType, is used to denote fields whose type is declared
elsewhere.

We also make a few other changes to support schema definition files that
produce something besides events.

At this point we don't attempt to actually generate the non-event types.
Consequently this code doesn't actually work.

Some identifiers are left with bad names, e.g. if they contain "event"
but no longer necessarily represent an event. That will be addressed in
a separate "rename all the things commit".

For obvious reasons we also had to bump the protocol repo submodule
reference to a commit that actually includes the schema definition
files. Unfortunately we also got another schema change that bumped
the version of the TERCC event.