Generate non-event struct types and make everything work
Since we want to be able to generate both event types and non-event
types (and their implementations will differ) we introduce a
schemaDefinitionRenderer interface whose implementing types will be
responsible for picking a template file and setting up all the template
variables. The schema discovery code is now responsible for picking the
right schemaDefinitionRenderer implementer based on the name of the
definition files.

One of the schemaDefinitionRenderer implementers is eventDefinitionFile,
which grabs the rendering code in generateEventFile from eventtypes.go.

Some types still have bad names but that'll get fixed in a follow-up.