Adapt to new YAML-based schema definitions (#40)
* internal/cmd/eventgen: Read YAML definition files instead of JSON schemas

We read the new YAML-based schema definition files and adapt the code to
deal with the JSON references that are used in those files. A new goType
type, goPredeclaredType, is used to denote fields whose type is declared
elsewhere.

We also make a few other changes to support schema definition files that
produce something besides events.

At this point we don't attempt to actually generate the non-event types.
Consequently this code doesn't actually work.

Some identifiers are left with bad names, e.g. if they contain "event"
but no longer necessarily represent an event. That will be addressed in
a separate "rename all the things commit".

For obvious reasons we also had to bump the protocol repo submodule
reference to a commit that actually includes the schema definition
files. Unfortunately we also got another schema change that bumped
the version of the TERCC event.

* Look up types of event's top-level members, don't hard code them

This required defining a custom function for event file template
expansions.

* Generate non-event struct types and make everything work

Since we want to be able to generate both event types and non-event
types (and their implementations will differ) we introduce a
schemaDefinitionRenderer interface whose implementing types will be
responsible for picking a template file and setting up all the template
variables. The schema discovery code is now responsible for picking the
right schemaDefinitionRenderer implementer based on the name of the
definition files.

One of the schemaDefinitionRenderer implementers is eventDefinitionFile,
which grabs the rendering code in generateEventFile from eventtypes.go.

Some types still have bad names but that'll get fixed in a follow-up.

* Rename all the things

To avoid renaming identifiers while code was being refactored we've
saved most of the nomenclature adjustments to this commit. There are no
functional changes in this commit.

* Remove unused leftover template variable

* Moar renaming (no functional changes)

* linkslice.go: Mark internal dep as first-party