editiongen: Drop go-git in favor of calling Git to read repo files
Using github.com/go-git/go-git to list the files in the repository
rather than just running "git ls-tree" doesn't make life easier for
us but adds a bunch of dependencies that we'd rather be without.
What's worse is that editiongen isn't in a module of its own so the
extra dependencies spill into our dependents.