Add support for experimental event versions
Nowadays, the protocol supports experimental (pre-1.0.0) versions,
with EiffelArtifactDeployedEvent being the first event of that kind.
The point of those versions is to allow experimentation and non
backwards compatible changes without major version bumps.

For this SDK, this means that all pre-1.0.0 versions need to get
their own struct types since there could be arbitrary changes between,
say, 0.1.0 and 0.2.0, and the only way to guarantee that we can
support all such versions is to use distinct types for all of them.

Since we're updating the protocol submodule gitlink in this commit
(to get hold of EiffelArtifactDeployedEvent) we're also getting a few
other changes, effectively making us cover nearly the whole intended
scope of the Orizaba edition.