Add schema-based event validation (#71)
Add a "validator" subpackage where types and functions for event
validation will be kept. An extendable model was chosen, where a
Set instance tracks one or more instances of Validator that each
know how to validate some aspect of an event.  For now we supply
one such implementation, SchemaValidator. In a similar way, that
type uses one or more SchemaLocator instances to locate schemas.
We include to such implementations; one that uses the compiled-in
schema files, and one that fetches schemas based on the event's
meta.schemaUri value.

To include the schema files we use the embed package, but the
original schema files in the submodule won't be available when
this package is imported by someone else so we have to synchronize
them into this git. To make sure we don't get out of sync, this
synchronization is done with a //go:generate invocation that
references a new binary that just wraps rsync. This introduces
rsync as a build dependency for developers of this package, but
that's acceptable. The makefile receives a slight patch so it not
only detects dirty files but also untracked ones (so that we detect
files that should've been under source control but aren't).