Add signature package for signing events and verifying signatures (#91)
The new package basically contains two types, Signer and Verifier,
which, unsurprisingly, sign events and verifies signatures of existing
events.

Signer instances are configured with an identity and a private key
and signs events into byte slices. Verifier instance require you
to pass something that implements the PublicKeyLocator interface.
That interface looks up which public key(s) can be used to verify
the signature of an event with a given meta.security.authorIdentity.
Because public key lookups are expected to be application-dependent
we don't include a type that implements PublicKeyLocator, but that
might change over time once we understand typical usage patterns.

Since we wanted to use the errors.Join function, introduced in
Go 1.20, we stepped the Go version requirement in go.mod to 1.21.
That's the currently oldest supported version so it's a totally
reasonable requirement.

The compiler upgrade triggered a new linter violation because of
the deprecation of strings.Title in Go 1.18, so we had to address
that too.