signature: Add CapabilityTeller interface (#94)
This new interface, currently with a single method that indicates
whether the event can be signed, is used to return an error if
the event given to Signer.Sign is too old to support signing.
All events implement the CapabilityTeller interface. In the future
we might want to add additional methods for expressing other
capabilities of events, e.g. if they support meta.source.domainId.